Implemented binance adapter

+++ api/src/main/java/org/nl/bot/api/beans/Order.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 
 public interface Order {
 
-    int getLots();
+    double getLots();
     @Nonnull
     Operation getOperation();
     @Nonnull
